Add CoveringIndexTrait and CoveringIndexConfigTrait
